modificando ubicacion de algunos componentes
