+++ src/main/java/org/tallerjava/moduloSeguridad/infraestructura/rateLimiter/RateLimitFilter.java
@@ -1,4 +1,4 @@
-package org.tallerjava.servicioExterno.rateLimiter;
+package org.tallerjava.moduloSeguridad.infraestructura.rateLimiter;
 
 import jakarta.ws.rs.container.ContainerRequestContext;
 import jakarta.ws.rs.container.ContainerRequestFilter;
