+++ src/main/java/org/tallerjava/moduloTransferencia/interfase/remota/ClienteHttpTransferencia.java
@@ -1,6 +1,4 @@
-package org.tallerjava.servicioExterno.interfase.remota;
-
-import org.tallerjava.servicioExterno.datatypes.DTONotificacionTransferencia;
+package org.tallerjava.moduloTransferencia.interfase.remota;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.ws.rs.client.Client;
@@ -11,40 +9,6 @@ import jakarta.ws.rs.core.Response;
 
 @ApplicationScoped
 public class ClienteHttpTransferencia {
-
-    private static final String ENDPOINT_TRANSFERENCIA = "http://localhost:8080/TallerJakartaEEPasarelaPagos/api/transferencia/notificacion";
-    private static final String ENDPOINT_COMPRA = "http://localhost:8080/TallerJakartaEEPasarelaPagos/api/compra/notificacion";
-
-    public boolean enviarNotificacion(DTONotificacionTransferencia dto) {
-        try {
-            Client client = ClientBuilder.newClient();
-            Response response = client
-                    .target(ENDPOINT_TRANSFERENCIA)
-                    .request(MediaType.APPLICATION_JSON)
-                    .post(Entity.entity(dto, MediaType.APPLICATION_JSON));
-
-            return response.getStatus() == 200;
-        } catch (Exception e) {
-            e.printStackTrace();
-            return false;
-        }
-    }
-
-    public boolean enviarRespuestaProcesoPago(boolean exito) {
-        try {
-            Client client = ClientBuilder.newClient();
-            Response response = client
-                .target(ENDPOINT_COMPRA)
-                .request(MediaType.APPLICATION_JSON)
-                .post(Entity.entity(exito, MediaType.APPLICATION_JSON));
-
-            return response.getStatus() == 200;
-        } catch (Exception e) {
-            e.printStackTrace();
-            return false;
-        }
-    }
-
     public boolean notificarBancoSoapHttp(String numeroCuenta, double monto, String codigoTransaccion) {
         try {
             String soapEndpointUrl = "http://localhost:8080/TallerJakartaEEPasarelaPagos/NotificacionBancoService";
